Exercises/requirements.txt
apache-airflow==2.7.1
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2 